FIX[Formacion]: Fix error name in director
