+++ packages/api-gateway/src/drivers/http/routes/planteles/schema/formacionDirectores/find-one.formacion-director.schema.js
@@ -2,29 +2,26 @@ const { formacion } = require('../../../instituciones/schema/formacionesRectores
 const { formacionDirector } = require('./properties/formacionDirector');
 const { responseProperties } = require('../properties/responseProperties');
 
-const findAllFormacionDirectorSchema = {
+const findOneFormacionDirectorSchema = {
   tags: ['Formacion'],
-  description: 'Get all the formations of a rector in institutions.',
+  description: 'Get details of a formation of a director in institutions.',
   params: {
     type: 'object',
     properties: {
-      ...formacionDirector.directorId,
+      ...formacionDirector,
     },
-    required: ['directorId'],
+    required: ['directorId', 'formacionId'],
   },
   response: {
     200: {
       type: 'object',
       properties: {
         data: {
-          type: 'array',
-          items: {
-            type: 'object',
-            properties: {
-              id: { type: 'integer' },
-              ...formacion,
-              ...responseProperties,
-            },
+          type: 'object',
+          properties: {
+            id: { type: 'integer' },
+            ...formacion,
+            ...responseProperties,
           },
         },
       },
@@ -32,4 +29,4 @@ const findAllFormacionDirectorSchema = {
   },
 };
 
-module.exports = { findAllFormacionDirectorSchema };
+module.exports = { findOneFormacionDirectorSchema };
