+++ packages/api-gateway/src/drivers/http/routes/planteles/schema/formacionDirectores/index.js
@@ -1,7 +1,7 @@
-const { createFormacionDirectorSchema } = require('./create.formacion-rector.schema');
-const { updateFormacionDirectorSchema } = require('./update.formacion-rector.schema');
-const { findAllFormacionDirectorSchema } = require('./find-all.formacion-rector.schema');
-const { findOneFormacionDirectorSchema } = require('./find-one.formacion-rector.schema');
+const { createFormacionDirectorSchema } = require('./create.formacion-director.schema');
+const { updateFormacionDirectorSchema } = require('./update.formacion-director.schema');
+const { findAllFormacionDirectorSchema } = require('./find-all.formacion-director.schema');
+const { findOneFormacionDirectorSchema } = require('./find-one.formacion-director.schema');
 
 module.exports = {
   createFormacionDirectorSchema,
